% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/br-metadata-getters.R
\name{br_read_crawler_metadata}
\alias{br_read_crawler_metadata}
\title{Read current prices table}
\usage{
br_read_crawler_metadata()
}
\value{

}
\description{
gets you a html table that you can use to extract current prices
(by simply applying html_table) or companies names required to run a crawler
(names in url are different both from ticker and the company name here)
}
