% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawler-functions-and-utils.R
\name{br_read_crawler_results}
\alias{br_read_crawler_results}
\title{Read crawler results}
\usage{
br_read_crawler_results(dir)
}
\arguments{
\item{dir}{a directory name where crawler results are stored}
}
\value{
a tibble with all results
}
\description{
Read crawler results
}
