% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawler-functions-and-utils.R
\name{br_run_crawler}
\alias{br_run_crawler}
\title{Run crawler}
\usage{
br_run_crawler(
  names_for_crawler,
  .f,
  dir = "crawler_results",
  sleep = TRUE,
  sleep_mean_delay = 10,
  ...
)
}
\arguments{
\item{names_for_crawler}{vector of names to iterate on. See examples.}

\item{.f}{a function to use for iteration. You can use all functions with a
prefix "bri_". As of now only bri_get_financial_report is available}

\item{dir}{a directory name where the crawler results should be stored.
Default to "crawler_results"}

\item{sleep}{TRUE/FALSE whether between crawler requests should be a pause.
Default to TRUE as I don't know if there is some anti-crawler filter here
at play}

\item{sleep_mean_delay}{an average time (in seconds) that the crawler should
wait to get another company's data. Pause times are randomly distributed.}
}
\value{
nothing xd
}
\description{
Run crawler
}
\examples{
data(crawler_metadata)
vector_of_names <- crawler_metadata$name_for_crawler
br_run_crawler(
  vector_of_names,
  bri_get_pnl,
  dir = "gpw_pnls"
)
# ... here's where you wait
all_reports_data <- br_read_crawler_data("gpw_pnls")
}
