% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawler-functions-and-utils.R
\name{construct_iteration_table}
\alias{construct_iteration_table}
\title{Create iteration table}
\usage{
construct_iteration_table(companies_list)
}
\arguments{
\item{companies_list}{you can either get it from br_read_crawler_metadata()
or crawler_metadata tibble saved in a package. See example:}
}
\value{
tibble
}
\description{
you can probably download the data without it just iterating on
the vector of companies names but in case there's some anti-crawler filter
applied it seemed wise to create a random order iterator
}
\examples{
data(crawler_metadata)
iteration_table <- construct_iteration_table(crawler_metadata$name_for_crawler)
}
